Servant.OAuth2.Examples.Cookies

Description

This example follows the Servant.OAuth2.Examples.Simple example very closely, but this time we use a configuration that let's enables us to set a cookie, and then redirect to the homepage.

Moreover, we set things up so that we can read that cookie on any page, to determine if the current visitor is logged in.

We will assume you have read the Simple example, and mostly spend our time explaining what is different.

Documentation

type OAuth2Result = '[WithStatus 303 RedirectWithCookie]

This time our result type is a set of headers that both redirects, and sets a particular cookie value. The cookie will, here, contain simply the result of the oauth2 workflow; i.e. the users email.

Since: 0.1.0.0

optionalUserAuthHandler :: Key -> AuthHandler Request (Maybe Text)

This is the corresponding handler for the above instance. Our implementation is very simple, we just call getSessionIdFromCookie, which is provided by the Servant.OAuth2 library itself; this decodes a previously-encoded value from the cookie, by the corresponding function buildSessionCookie, which we will later use through the simpleCookieOAuth2Settings function.

Since: 0.1.0.0

data Routes mode

As last time, we have our routes; the main change is the inclusion of the AuthProtect tag on the home route, that let's us bring a potential user into scope for that page.

Since: 0.1.0.0

Constructors

Routes

Fields

home :: mode :- (AuthProtect "optional-cookie" :> Get '[HTML] Html)

auth :: mode :- (AuthProtect Github :> ("auth" :> ("github" :> NamedRoutes (OAuth2Routes OAuth2Result))))

mkSettings :: Key -> OAuthConfig -> OAuth2Settings Handler Github OAuth2Result

Again, we have settings, but this time, instead of using the defaultOAuth2Settings, we use the simpleCookieOAuth2Settings function to get default behaviour that, upon successful completion of the oauth2 flow, builds a cookie with a session id &mdash; in this case just the ident of the user &mdash; and then redirects the browser to the homepage.

Since: 0.1.0.0

server :: OAuthConfig -> OAuth2Settings Handler Github OAuth2Result -> Routes (AsServerT Handler)

Now we can have a simple server implementation, but this time we can check if the user us logged in by looking at the first parameter to the home function; i.e. if it's Nothing then we're not logged in, otherwise we are! Very convenient.

Since: 0.1.0.0

main :: IO ()

Our entrypoint; the only addition here is that we need to obtain a Key to do our cookie encryption/decryption; and we again need to build up our context with our Github-based oauth2AuthHandler and our own custom one, optionalUserAuthHandler, to decode the cookie.

Since: 0.1.0.0